+++ nilearn/input_data/tests/test_nifti_labels_masker.py
@@ -14,7 +14,6 @@ from nilearn.input_data.nifti_labels_masker import NiftiLabelsMasker
 from nilearn._utils import testing, as_ndarray
 from nilearn._utils.exceptions import DimensionError
 from nilearn._utils.testing import assert_less
-from nilearn._utils.compat import get_affine
 
 
 def generate_random_img(shape, length=1, affine=np.eye(4),
@@ -102,8 +101,7 @@ def test_nifti_labels_masker():
     # Call inverse transform (smoke test)
     fmri11_img_r = masker11.inverse_transform(signals11)
     assert_equal(fmri11_img_r.shape, fmri11_img.shape)
-    np.testing.assert_almost_equal(get_affine(fmri11_img_r),
-                                   get_affine(fmri11_img))
+    np.testing.assert_almost_equal(fmri11_img_r.affine, fmri11_img.affine)
 
 
 def test_nifti_labels_masker_with_nans_and_infs():
@@ -160,20 +158,20 @@ def test_nifti_labels_masker_resampling():
                                resampling_target="labels")
 
     masker.fit()
-    np.testing.assert_almost_equal(get_affine(masker.labels_img_),
-                                   get_affine(labels33_img))
+    np.testing.assert_almost_equal(masker.labels_img_.affine,
+                                   labels33_img.affine)
     assert_equal(masker.labels_img_.shape, labels33_img.shape)
 
-    np.testing.assert_almost_equal(get_affine(masker.mask_img_),
-                                   get_affine(masker.labels_img_))
+    np.testing.assert_almost_equal(masker.mask_img_.affine,
+                                   masker.labels_img_.affine)
     assert_equal(masker.mask_img_.shape, masker.labels_img_.shape[:3])
 
     transformed = masker.transform(fmri11_img)
     assert_equal(transformed.shape, (length, n_regions))
 
     fmri11_img_r = masker.inverse_transform(transformed)
-    np.testing.assert_almost_equal(get_affine(fmri11_img_r),
-                                   get_affine(masker.labels_img_))
+    np.testing.assert_almost_equal(fmri11_img_r.affine,
+                                   masker.labels_img_.affine)
     assert_equal(fmri11_img_r.shape,
                  (masker.labels_img_.shape[:3] + (length,)))
 
@@ -200,12 +198,12 @@ def test_nifti_labels_masker_resampling():
                                resampling_target="labels")
 
     masker.fit()
-    np.testing.assert_almost_equal(get_affine(masker.labels_img_),
-                                   get_affine(labels33_img))
+    np.testing.assert_almost_equal(masker.labels_img_.affine,
+                                   labels33_img.affine)
     assert_equal(masker.labels_img_.shape, labels33_img.shape)
 
-    np.testing.assert_almost_equal(get_affine(masker.mask_img_),
-                                   get_affine(masker.labels_img_))
+    np.testing.assert_almost_equal(masker.mask_img_.affine,
+                                   masker.labels_img_.affine)
     assert_equal(masker.mask_img_.shape, masker.labels_img_.shape[:3])
 
     uniq_labels = np.unique(masker.labels_img_.get_data())
@@ -218,8 +216,8 @@ def test_nifti_labels_masker_resampling():
     assert_less((transformed.var(axis=0) == 0).sum(), n_regions)
 
     fmri11_img_r = masker.inverse_transform(transformed)
-    np.testing.assert_almost_equal(get_affine(fmri11_img_r),
-                                   get_affine(masker.labels_img_))
+    np.testing.assert_almost_equal(fmri11_img_r.affine,
+                                   masker.labels_img_.affine)
     assert_equal(fmri11_img_r.shape,
                  (masker.labels_img_.shape[:3] + (length,)))
 
@@ -234,9 +232,8 @@ def test_nifti_labels_masker_resampling():
     masker = NiftiLabelsMasker(labels33_img, mask_img=mask22_img)
 
     masker.fit_transform(fmri22_img)
-    np.testing.assert_array_equal(
-        get_affine(masker._resampled_labels_img_),
-        affine2)
+    np.testing.assert_array_equal(masker._resampled_labels_img_.affine,
+                                  affine2)
 
     # Test with filenames
     with testing.write_tmp_imgs(fmri22_img) as filename: